Removed dependency for webdriver-manager, separated tests into different files, changed implicit wait, reduced test data
