+++ config/web_ui_config.yaml
@@ -1,4 +1,4 @@
 log_file_name: "web_ui_tests.log"
 executor: "http://selenium-hub:4444"
-browser: chrome # chrome, firefox or edge
-implicit_wait_time: 5
+browser: firefox # chrome, firefox or edge
+implicit_wait_time: 3
