+++ pyproject.toml
@@ -10,7 +10,6 @@ package-mode = false
 python = "^3.12"
 pytest = "^8.3.4"
 requests = "^2.32.3"
-webdriver-manager = "^4.0.2"
 selenium = "4.27.1"
 pyyaml = "^6.0.2"
 pytest-custom-outputs = "^1.1.1"
