+++ test_data/web_ui/entire_flow.yaml
@@ -3,13 +3,3 @@
   first_name: Desislava
   last_name: Dakova
   postal_code: 1220
-- username: visual_user
-  password: secret_sauce
-  first_name: Desislava
-  last_name: Dakova
-  postal_code: 1220
-- username: performance_glitch_user
-  password: secret_sauce
-  first_name: Desislava
-  last_name: Dakova
-  postal_code: 1220
